Fix big decimal (#35)
* some changes

1. replace integer instead of bigdecimal
2. add github actions
3. add linter
4. config unit test files correctly

* rename github action file

* change github command verify

* Revert "change github command verify"

This reverts commit 78877eb947b14c454c55a223f1b42d6211a8dcd1.

* add checkstyle step

* update version

add 3.0.2-SNAPSHOT

+++ src/main/java/conekta/io/model/impl/Order.java
@@ -5,7 +5,6 @@ import conekta.io.model.PaginatedConektaObject;
 import conekta.io.model.submodel.*;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 /**
  * Frase corta descriptiva
@@ -24,8 +23,8 @@ public class Order extends ConektaObject {
      * Comportamiento en caso de que sea null(si aplica)
      */
     private Boolean livemode;
-    private BigDecimal amount;
-    private BigDecimal amountRefunded;
+    private Integer amount;
+    private Integer amountRefunded;
     private String paymentStatus;
     private String currency;
     private Integer createdAt;
